langGraph/requirements.txt
langchain
langchain_community
langchain_openai
langgraph
python-dotenv
ipykernel